goto

+++ Debugger/Debugger.h
@@ -2,7 +2,7 @@
 #include <stdio.h>
 
 typedef struct fdebugger {
-    size_t (*log)(const char*, ...);
+    int (*log)(const char*, ...);
     void (*start)();
     void (*end)();
     void (*setMemory)(char*, size_t);
